journal: avoid race between in-flight notifications and flush
If an async callback for a safely committed event was in-flight,
it could race with the flush of the journal.  This would result
in the flush callback completing before the notifications for
safe events.

Fixes: #14434
